requirements.txt
alembic==1.10.3
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-Migrate==4.0.4
Flask-MySQL==1.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
pika==1.3.1
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wonderwords==2.2.0
zipp==3.15.0
